chore(evmcrispr): rename helpers folder to utils
