+++ packages/evmcrispr/src/Connector.ts
@@ -1,7 +1,7 @@
 import fetch from 'isomorphic-fetch';
 
-import type { GraphQLBody } from './helpers';
-import { ORGANIZATION_APPS, REPO, parseApp, parseRepo } from './helpers';
+import type { GraphQLBody } from './utils';
+import { ORGANIZATION_APPS, REPO, parseApp, parseRepo } from './utils';
 import { ErrorException, ErrorNotFound } from './errors';
 import type { Address, ParsedApp, Repo } from './types';
