+++ packages/evmcrispr/src/utils/apps.ts
