+++ packages/evmcrispr/src/utils/evmscripts.ts
