+++ packages/evmcrispr/src/utils/interfaces.ts
