+++ packages/evmcrispr/src/utils/queries.ts
