+++ packages/evmcrispr/test/test-helpers/expects.ts
@@ -3,7 +3,7 @@ import { utils } from 'ethers';
 import { multihash } from 'is-ipfs';
 
 import { ErrorInvalid } from '../../src';
-import { parseContentUri } from '../../src/helpers';
+import { parseContentUri } from '../../src/utils';
 import type { AragonArtifact, ParsedApp } from '../../src/types';
 
 const { isAddress } = utils;
